requirements.txt
aiogram
python-dotenv
asyncpg
sqlalchemy